New file for command utilities
